hello-world-1/requirements.txt
Flask==2.0.2
requests==2.26.0
ptvsd==4.3.2 # Required for debugging.
SQLAlchemy==1.4.29
PyMySQL==1.0.2
gunicorn==20.1.0
google-cloud-logging==2.7.0